Move file-upload-dialog component into its own folder
