+++ src/app/components/file-upload/file-upload-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './file-upload-dialog.component';
